requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
click==8.2.1
fastapi==0.116.0
frozenlist==1.7.0
h11==0.16.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
slack_bolt==1.23.0
slack_sdk==3.36.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
uvicorn==0.35.0
yarl==1.20.1
